add notebook for paper live kraken
