+++ src/qubx/connectors/ccxt/exchanges/binance/exchange.py
@@ -3,7 +3,6 @@ from typing import Dict, List, cast
 import ccxt.pro as cxp
 from ccxt.async_support.base.ws.cache import ArrayCache, ArrayCacheByTimestamp
 from ccxt.async_support.base.ws.client import Client
-from ccxt.async_support.base.ws.future import Future
 from ccxt.base.errors import ArgumentsRequired, BadRequest, InsufficientFunds, NotSupported
 from ccxt.base.precise import Precise
 from ccxt.base.types import (
@@ -16,61 +15,17 @@ from ccxt.base.types import (
     Strings,
 )
 
+from ..base import CcxtFuturePatchMixin
 
-class BinanceQV(cxp.binance):
+
+class BinanceQV(CcxtFuturePatchMixin, cxp.binance):
     """
     Extended binance exchange to provide quote asset volumes support and fix bulk subscription issues
     """
 
     def __init__(self, config={}):
         super().__init__(config)
-        # Patch the Future.race method to prevent InvalidStateError
-        self._patch_future_race()
 
-    def _patch_future_race(self):
-        """Patch CCXT Future.race to prevent InvalidStateError race condition."""
-        import asyncio
-
-        original_race = Future.race
-
-        @classmethod
-        def patched_race(cls, futures):
-            future = Future()
-            coro = asyncio.wait(futures, return_when=asyncio.FIRST_COMPLETED)
-            task = asyncio.create_task(coro)
-
-            def callback(done):
-                complete, _ = done.result()
-                # Check for exceptions
-                exceptions = []
-                cancelled = False
-                for f in complete:
-                    if f.cancelled():
-                        cancelled = True
-                    else:
-                        err = f.exception()
-                        if err:
-                            exceptions.append(err)
-
-                # Only set result/exception if future is not already done (prevents InvalidStateError)
-                if future.done():
-                    return
-
-                # If any exceptions return with first exception
-                if len(exceptions) > 0:
-                    future.set_exception(exceptions[0])
-                # Else return first result
-                elif cancelled:
-                    future.cancel()
-                else:
-                    first_result = list(complete)[0].result()
-                    future.set_result(first_result)
-
-            task.add_done_callback(callback)
-            return future
-
-        # Apply the patch
-        Future.race = patched_race
 
     def describe(self):
         """
