+++ src/qubx/connectors/ccxt/exchanges/kraken/kraken.py
@@ -1,7 +1,9 @@
 import ccxt.pro as cxp
 
+from ..base import CcxtFuturePatchMixin
 
-class CustomKrakenFutures(cxp.krakenfutures):
+
+class CustomKrakenFutures(CcxtFuturePatchMixin, cxp.krakenfutures):
     def describe(self):
         # we disable watchBidsAsks because it does not work well
         return self.deep_extend(
